requirements.txt
python3
tkinter
PyNaCl (Python Networking and Cryptograpgy libary)